Created "modules" folder and moved all modules into it
