+++ apps/api/src/modules/auth/auth.controller.ts
@@ -4,8 +4,8 @@ import { ConfigType } from '@nestjs/config';
 import { ApiBody } from '@nestjs/swagger';
 import { User as PrismaUser } from '@prisma/client';
 import { Request, Response } from 'express';
-import { globalConfig } from '../config';
-import { messageWhenCalled } from '../utils/decorators';
+import { globalConfig } from '../../config';
+import { messageWhenCalled } from '../../utils/decorators';
 import { AuthService } from './auth.service';
 import { LoginDto } from './dtos';
 import { GoogleAuthGuard } from './google-auth.guard';
