+++ apps/api/src/modules/auth/auth.module.ts
