+++ apps/api/src/modules/auth/local.strategy.ts
