+++ apps/api/src/modules/users/users.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { PrismaService } from '../prisma/prisma.service';
+import { PrismaService } from '../../prisma/prisma.service';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
