+++ apps/api/src/modules/users/users.service.ts
@@ -2,8 +2,8 @@ import { HttpException, HttpStatus, Inject, Injectable } from '@nestjs/common';
 import { ConfigType } from '@nestjs/config';
 import { Prisma } from '@prisma/client';
 import bcrypt from 'bcrypt';
-import { globalConfig } from '../config/config';
-import { PrismaService } from '../prisma/prisma.service';
+import { globalConfig } from '../../config/config';
+import { PrismaService } from '../../prisma/prisma.service';
 
 export type User = {
 	id: number;
